chore(persistence): move package
